+++ yody-be/src/main/java/org/example/yodybe/service/SizeServiceImpl.java
@@ -1,8 +1,11 @@
 package org.example.yodybe.service;
 
+import jakarta.transaction.Transactional;
 import org.example.yodybe.dto.SizeDto;
+import org.example.yodybe.entity.Product;
 import org.example.yodybe.entity.Size;
 import org.example.yodybe.form.SizeForm;
+import org.example.yodybe.repositoties.ProductRepository;
 import org.example.yodybe.repositoties.SizeRepository;
 import org.example.yodybe.utils.BaseResponse;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,6 +19,8 @@ import java.util.stream.Collectors;
 public class SizeServiceImpl implements SizeService {
     @Autowired
     private SizeRepository sizeRepo;
+    @Autowired
+    private ProductRepository productRepository;
 
     @Override
     public BaseResponse getAllSizes() {
@@ -42,17 +47,25 @@ public class SizeServiceImpl implements SizeService {
     public BaseResponse saveSize(SizeForm color) {
         try {
             Size entity = mapToEntity(color);
-            sizeRepo.save(entity);
-            return new BaseResponse("Lưu thành công", true, 200);
+            Size res = sizeRepo.save(entity);
+            return new BaseResponse("Lưu thành công", res, 200);
         } catch (Exception e) {
-            return new BaseResponse("Lưu thất bại", true, 400);
+            return new BaseResponse("Lưu thất bại", false, 400);
         }
     }
 
+    @Transactional
     @Override
-    public BaseResponse deleteSize(Long id) {
+    public BaseResponse deleteSize(List<Long> ids) {
         try {
-            sizeRepo.deleteById(id);
+            for (Long id : ids) {
+                Optional<Size> size = sizeRepo.findById(id);
+                List<Product>  products = productRepository.findBySizes(size.get());
+                if (!products.isEmpty()) {
+                    return new BaseResponse("Không thể xóa màu ["+size.get().getName()+"] đang có sản phẩm liên quan", true, 400);
+                }
+                sizeRepo.deleteById(id);
+            }
             return new BaseResponse("Xóa thành công", true, 200);
         } catch (Exception e) {
             return new BaseResponse("Xóa thất bại", true, 200);
@@ -67,6 +80,7 @@ public class SizeServiceImpl implements SizeService {
                 return new BaseResponse("Không tìm thấy màu " + clSizeForm.getName(), false, 4000);
             }
             size.get().setName(clSizeForm.getName());
+            sizeRepo.save(size.get());
             return new BaseResponse("Cập nhật thành công", true, 200);
         } catch (Exception e) {
             return new BaseResponse("Cập nhật thất bại ", true, 200);
